e2e test adjustments
