+++ e2e/test/scenarios/data-model/data-model-shared-4.cy.spec.ts
@@ -6,24 +6,21 @@ import {
 import { SAMPLE_DATABASE } from "e2e/support/cypress_sample_database";
 
 const { H } = cy;
-const { TablePicker, TableSection, FieldSection, PreviewSection, Shared } =
+const { TablePicker, TableSection, FieldSection, PreviewSection } =
   cy.H.DataModel;
 
 const { ORDERS_ID, ORDERS } = SAMPLE_DATABASE;
-const { visitArea } = Shared;
 
-const areas: ("admin" | "data studio")[] = ["admin", "data studio"];
-type Area = (typeof areas)[number];
+const visit = H.DataModel.visitDataStudio;
 
-describe.each<Area>(areas)("data model > %s", (area: Area) => {
-  const visit = visitArea(area);
-
+describe("data model", () => {
   beforeEach(() => {
     H.restore();
     H.resetSnowplow();
     cy.signInAsAdmin();
     H.activateToken("bleeding-edge");
 
+    cy.intercept("GET", "/api/database").as("databases");
     cy.intercept("GET", "/api/database/*/schemas?*").as("schemas");
     cy.intercept("GET", "/api/table/*/query_metadata*").as("metadata");
     cy.intercept("GET", "/api/database/*/schema/*").as("schema");
@@ -37,16 +34,6 @@ describe.each<Area>(areas)("data model > %s", (area: Area) => {
     cy.intercept("POST", "/api/field/*/dimension").as("updateFieldDimension");
     cy.intercept("PUT", "/api/table").as("updateTables");
     cy.intercept("PUT", "/api/table/*").as("updateTable");
-
-    if (area === "admin") {
-      cy.intercept("GET", "/api/database?*").as("databases");
-      cy.intercept("GET", "/api/field/*/values").as("fieldValues");
-      cy.intercept("PUT", "/api/table/*").as("updateTable");
-    }
-
-    if (area === "data studio") {
-      cy.intercept("GET", "/api/database").as("databases");
-    }
   });
 
   describe("Error handling", { tags: "@external" }, () => {
@@ -66,15 +53,9 @@ describe.each<Area>(areas)("data model > %s", (area: Area) => {
       cy.intercept("POST", "/api/field/*/values", error);
       cy.intercept("POST", "/api/field/*/dimension", error);
       cy.intercept("PUT", "/api/table/*", error);
-      if (area === "admin") {
-        cy.intercept("POST", "/api/table/*/sync_schema", error);
-        cy.intercept("POST", "/api/table/*/rescan_values", error);
-        cy.intercept("POST", "/api/table/*/discard_values", error);
-      } else {
-        cy.intercept("POST", "/api/data-studio/table/sync-schema", error);
-        cy.intercept("POST", "/api/data-studio/table/rescan-values", error);
-        cy.intercept("POST", "/api/data-studio/table/discard-values", error);
-      }
+      cy.intercept("POST", "/api/data-studio/table/sync-schema", error);
+      cy.intercept("POST", "/api/data-studio/table/rescan-values", error);
+      cy.intercept("POST", "/api/data-studio/table/discard-values", error);
     });
 
     it("shows toast errors and preview errors", () => {
